Simplify package
Previously every Each() goroutine iterated over a single list needing
to keep track of the current position and be notified when it is
appended to. This was more complicated then necessary and was issues
with race conditions.

To simplify this a pool of chan's are appended to each time Add(...)
is called. To ensure later call's to Each() recieve all messages a
backlog of each message is also contained on a shared slice. This
backlog is returned on the channel returned from Each() before
Add(...) starts sending new items directly to that channel.